pyproject.toml
[tool.poetry]
name = "template-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
PyYAML = "^6.0.1"
fastapi = "^0.111.1"
uvicorn = { extras = ["standard"], version = "^0.30.3" }
mangum = "^0.17.0"
pydantic = "^2.8.2"
starlette = "^0.37.2"
pytz = "^2022.7.1"
boto3 = "^1.34.149"
python-dotenv = "^1.0.1"
asgi-correlation-id = "^4.3.1"
SQLAlchemy = "^2.0.31"
dependency-injector = "^4.41.0"
psycopg2-binary = "2.9.9"
celery = "5.4.0"
flower = "2.0.0"
pycurl = "7.45.3"
redis = "^5.0.8"
passlib = "1.7.4"
pyjwt = "2.9.0"
bcrypt= "4.2.0"


[tool.poetry.dev-dependencies]
pylint = "3.2.6"
coloredlogs = "14.0.0"
colorlog = "4.2.1"
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
